requirements.txt
flask
flask-cors
python-dotenv
openai
PyPDF2
pytesseract
pdf2image
python-docx 